Added my Higher Lower Project within JavaScript
This project utilizes basic CSS, HTML, and intermediate JavaScript code that creates a webpage that simulates a number guessing game. On load, the page asks for a positive integer above 1. The user will enter a guess into a text input and click a button to submit the guess. It will notify the user when they guessed a correct answer.